Use multiple connections and chunking
Use multiple connections to send pixels to the server. For each pixel, a
connections is selected at random. To direct a connection to send a
pixel, channels are used so that Pixelflut commands are only sent one at
a time.

Furthermore, the image is now split into evenly-sized chunks, and pixels
are scheduled to be sent from each one concurrently.

Closes #1.